Add additional validation

+++ src/Controllers/MessagesController.cs
@@ -29,7 +29,7 @@ namespace ChatApi.Controllers
             var message = collection.Find(m => m.Id == id).FirstOrDefault();
 
             if (message == null)
-                return NotFound();
+                return NotFound("Message not found");
 
             return Json(message);
         }
@@ -38,6 +38,11 @@ namespace ChatApi.Controllers
         [Route("~/api/users/{userId}/sentmessages")]
         public IActionResult GetAllFromUser(string userId, int limit)
         {
+            var uCollection = CollectionManager.GetUserCollection();
+            var exists = uCollection.Find(u => u.Id == userId).FirstOrDefault();
+            if (exists == null)
+                return NotFound("User not found");
+
             var collection = CollectionManager.GetMessageCollection();
             var messages = collection.Find(m => m.SenderId == userId).Limit(limit).ToList();
             return Json(messages);
@@ -63,6 +68,11 @@ namespace ChatApi.Controllers
         public IActionResult Delete(string id)
         {
             var collection = CollectionManager.GetMessageCollection();
+
+            var exists = collection.Find(m => m.Id == id).FirstOrDefault();
+            if (exists == null)
+                return NotFound("Message not found");
+
             collection.DeleteOne(m => m.Id == id);
             return Ok();
         }
